Fix type mismatch issue
1. refs #392

+++ tgs/src/main/cpp/tgs/BigContainers/BigMapStxxl.h
@@ -99,7 +99,7 @@ public:
       return 0;
     }
     typename MapType::const_iterator it = _map->find(k);
-    return it == _map->end() ? 0 : 1;
+    return it == _map->end() ? (size_t)0 : (size_t)1;
   }
 
   void insert(const K& k, const V& v) { _bloom->insert(k); _map->insert(std::pair<K, V>(k, v)); }
